+++ src/commands/mod.rs
@@ -6,6 +6,7 @@ mod lint;
 mod pitstop;
 mod post_edit;
 mod precommit;
+mod update_tools;
 
 pub use ci::ci;
 pub use fix::fix;
@@ -15,3 +16,4 @@ pub use lint::lint;
 pub use pitstop::pitstop;
 pub use post_edit::post_edit;
 pub use precommit::precommit;
+pub use update_tools::update_tools;